Remove unused symlink option
Unfortunately the behaviour here was to set SYMLINK to "" in the case
that symlinks are not requested, which leads to "ln" being called like:

    ln "$SYMLINK" "$SOURCE" "$TARGET"

With $SYMLINK being an empty string you might expect that to be:

    ln "$SOURCE" "$TARGET"

but it actually does:

    ln "" "$SOURCE" "$TARGET"

which causes a failure to hardlink as it tries to use an empty path as
source.

Since we have no interest in symlinking, just remove the feature.