blockbackup: add an option to not print progress
By default, blockbackup prints a line of status output for each chunk
dealt with and if `--zip-after` is in use then another line is printed
when the compression happens. With the default chunk size of 1MiB this
will be result in 1,000 (or 2,000) lines of output per 1GiB of source
device.

Add a `--no-progress` option to avoid printing any of this.

Ultimately a better solution would be to detect a tty and do a progress
bar on a single line.